improved iPad version (image resizing), updated SVProgressHUD, did some integer casting

+++ Externals/DTCoreText/Core/Source/NSAttributedString+SmallCaps.m
@@ -21,7 +21,7 @@
 	}
 	
 	DTCoreTextFontDescriptor *smallerFontDesc = [fontDescriptor copy];
-	smallerFontDesc.pointSize *= 0.7;
+	smallerFontDesc.pointSize *= (float)0.7;
 	
 	CTFontRef smallerFont = [smallerFontDesc newMatchingFont];
